Improve file structure by moving `dispatch.c` to `src/oif` and headers to `/include`
