+++ CMakeLists.txt
@@ -7,11 +7,6 @@ if (CMAKE_BUILD_TYPE STREQUAL "Debug")
     set(CMAKE_C_FLAGS "${CMAKE_C_FLAGS} -O0")
 endif()
 
-add_library(oif_dispatch SHARED dispatch.c)
-target_compile_features(oif_dispatch PUBLIC c_std_11)
-target_include_directories(oif_dispatch PUBLIC .)
-target_include_directories(oif_dispatch PUBLIC include)
-
 add_subdirectory(src/oif)
 
 add_subdirectory(examples)
