+++ Makefile
@@ -10,7 +10,7 @@ endif
 all :
 	cmake -S . -B build -DCMAKE_VERBOSE_MAKEFILE:BOOL=TRUE -DCMAKE_BUILD_TYPE=Debug -DCMAKE_EXPORT_COMPILE_COMMANDS=ON && \
 	cmake --build build && \
-	cp build/liboif_dispatch.$(DSO_EXT) . && \
+	cp build/src/oif/liboif_dispatch.$(DSO_EXT) . && \
 	cp build/src/oif/backend_c/liboif_backend_c.$(DSO_EXT) .
 	cp build/src/oif/backend_c/liboif_backend_c_qeq.$(DSO_EXT) .
 	cp build/src/oif/backend_python/liboif_backend_python.$(DSO_EXT) .
