+++ include/oif/globals.h
