+++ src/oif/frontend_c/frontend_c.c
@@ -1,5 +1,5 @@
 #include <oif/api.h>
-#include "dispatch.h"
+#include <oif/dispatch.h>
 
 
 BackendHandle
