+++ src/oif/frontend_c/qeq.c
@@ -1,7 +1,8 @@
-#include "oif/api.h"
-#include <oif/frontend_c/qeq.h>
-#include <dispatch.h>
 #include <stdlib.h>
+
+#include <oif/api.h>
+#include <oif/dispatch.h>
+#include <oif/frontend_c/qeq.h>
 
 int oif_solve_qeq(
     BackendHandle bh, double a, double b, double c, OIFArrayF64 *roots
